Adds createTask
