+++ app/functions/annotateRunSessions.tsx
@@ -6,8 +6,8 @@ import type { Session } from "~/modules/sessions/sessions.types";
 import type { AnnotationSchemaItem, PromptVersion } from "~/modules/prompts/prompts.types";
 import getDocumentsAdapter from "~/modules/documents/helpers/getDocumentsAdapter";
 import type { Project } from "~/modules/projects/projects.types";
-import getQueue from "~/modules/queues/helpers/getQueue";
 import hasFeatureFlag from "~/modules/featureFlags/helpers/hasFeatureFlag";
+import createTask from "~/modules/queues/helpers/createTask";
 
 export default async function annotateRunSessions({ runId }: { runId: string }, context: { request: Request }) {
 
@@ -73,25 +73,17 @@ export default async function annotateRunSessions({ runId }: { runId: string }, 
       const hasWorkers = await hasFeatureFlag('HAS_WORKERS', context);
       if (run.data.annotationType === 'PER_UTTERANCE') {
         if (hasWorkers) {
-          const queue = getQueue('tasks');
-          const job = await queue.add('ANNOTATE_PER_SESSION', {
-            inputFile: `${inputDirectory}/${sessionModel.data._id}/${sessionModel.data.name}`,
-            outputFolder: `${outputDirectory}/${sessionModel.data._id}`,
-            prompt: { prompt: promptVersion.data.userPrompt, annotationSchema },
-            model: run.data.model,
-            team: project.data.team
-          }, {
-            removeOnComplete: {
-              age: 72 * 3600, // keep up to 1 hour
-              count: 2000, // keep up to 2000 jobs
-            },
-            removeOnFail: {
-              age: 72 * 3600,
-              count: 2000,
-            },
+          await createTask({
+            task: 'ANNOTATE_PER_UTTERANCE',
+            job: {
+              inputFile: `${inputDirectory}/${sessionModel.data._id}/${sessionModel.data.name}`,
+              outputFolder: `${outputDirectory}/${sessionModel.data._id}`,
+              prompt: { prompt: promptVersion.data.userPrompt, annotationSchema },
+              model: run.data.model,
+              team: project.data.team
+            }
           });
         } else {
-
           await annotatePerUtterance({
             body: {
               inputFile: `${inputDirectory}/${sessionModel.data._id}/${sessionModel.data.name}`,
@@ -104,23 +96,15 @@ export default async function annotateRunSessions({ runId }: { runId: string }, 
         }
       } else {
         if (hasWorkers) {
-
-          const queue = getQueue('tasks');
-          const job = await queue.add('ANNOTATE_PER_SESSION', {
-            inputFile: `${inputDirectory}/${sessionModel.data._id}/${sessionModel.data.name}`,
-            outputFolder: `${outputDirectory}/${sessionModel.data._id}`,
-            prompt: { prompt: promptVersion.data.userPrompt, annotationSchema },
-            model: run.data.model,
-            team: project.data.team
-          }, {
-            removeOnComplete: {
-              age: 72 * 3600, // keep up to 1 hour
-              count: 2000, // keep up to 2000 jobs
-            },
-            removeOnFail: {
-              age: 72 * 3600,
-              count: 2000,
-            },
+          await createTask({
+            task: 'ANNOTATE_PER_SESSION',
+            job: {
+              inputFile: `${inputDirectory}/${sessionModel.data._id}/${sessionModel.data.name}`,
+              outputFolder: `${outputDirectory}/${sessionModel.data._id}`,
+              prompt: { prompt: promptVersion.data.userPrompt, annotationSchema },
+              model: run.data.model,
+              team: project.data.team
+            }
           });
         } else {
           await annotatePerSession({
